Refactor in prep for DB update and delegate
