+++ Apps/Database/src/Models/Cacheable/HMACHash.cs
@@ -13,15 +13,30 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 // -------------------------------------------------------------------------
-namespace HealthGateway.Common.Models
+namespace HealthGateway.Database.Models.Cacheable
 {
-    using System;
+    using HealthGateway.Database.Constant;
 
     /// <summary>
-    /// Represents a generic hash object.
+    /// A hash of something.
     /// </summary>
-    public interface IHash
+    public class HMACHash : IHash
     {
+        /// <summary>
+        /// Gets or sets the pseudo random function that was used to generate this hash.
+        /// </summary>
+        public HashFunction PseudoRandomFunction { get; set; } = HashFunction.HMACSHA512;
+
+        /// <summary>
+        /// Gets or sets the iterations used to generate this hash.
+        /// </summary>
+        public int Iterations { get; set; }
+
+        /// <summary>
+        /// Gets or sets the base64 salt that was used in generating the hash.
+        /// </summary>
+        public string? Salt { get; set; }
+
         /// <summary>
         /// Gets or sets the base64 encoded hash.
         /// </summary>
