+++ Apps/Database/src/Models/Cacheable/IHash.cs
@@ -1,0 +1,30 @@
+﻿// -------------------------------------------------------------------------
+//  Copyright © 2019 Province of British Columbia
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+// -------------------------------------------------------------------------
+using System.Text.Json.Serialization;
+
+namespace HealthGateway.Database.Models.Cacheable
+{
+    /// <summary>
+    /// Represents a generic hash object.
+    /// </summary>
+    public interface IHash
+    {
+        /// <summary>
+        /// Gets or sets the base64 encoded hash.
+        /// </summary>
+        public string? Hash { get; set; }
+    }
+}
